pyproject.toml
# Build system configuration
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

# Project metadata
[project]
name = "devtoolbox"
version = "0.1.0"
description = "Developer toolbox with text processing capabilities"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"

# Core dependencies
dependencies = [
    "openai>=1.0.0",
    "jira>=3.5.0",
    "markdown>=3.4.0",
    "python-gitlab>=3.15.0",
    "pypandoc>=1.11.0",
    "requests>=2.28.0",
    "pypinyin>=0.48.0",
    "retry>=0.9.2",
    "Pillow>=9.4.0",
    "spacy>=3.0.0",
    "langchain>=0.0.1",
    "tiktoken>=0.0.1",
    "transformers>=4.0.0",
    "langdetect>=1.0.9",
]

# Package configuration
[tool.setuptools]
packages = ["devtoolbox"]

# Command line registration (currently commented out)
#[tool.setuptools.entry_points]
#console_scripts = [
#    "devtoolbox-hello = devtoolbox.cli:hello"
#]

# Optional dependencies for different use cases
[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=21.5b2",
]
docs = ["mkdocs>=1.4.2", "mkdocstrings>=0.21.2"]

[project.scripts]
devtoolbox-setup-nlp = "devtoolbox.cmd.setup_nlp:download_spacy_models"